+++ src/components/Groups/GroupDetail.vue
@@ -22,13 +22,17 @@
             slot="controls"
             slot-scope="data">
 
-            <font-awesome-icon
+            <b-button
+              v-b-tooltip.hover
               v-if="isGroupAdmin"
-              icon="trash"
-              size="1x"
-              style="cursor: pointer;"
-              @click="removeUser(data.item.username)"/>
-
+              title="Remove user from group"
+              class="trash-button"
+              @click="removeUser(data.item.username, uiUser)">
+              <font-awesome-icon
+                icon="trash"
+                size="1x"
+                style="cursor: pointer;"/>
+            </b-button>
           </template>
         </b-table>
 
@@ -61,6 +65,7 @@
 </template>
 
 <script>
+import {mapState} from 'vuex';
 import GroupUserAdd from "./GroupUserAdd.vue";
 import Help from "../Help.vue";
 import IconLink from "../IconLink.vue";
@@ -103,18 +108,28 @@ export default {
       }
     };
   },
-  computed: {
-    isGroupAdmin() {
+  computed: mapState({
+    uiUser: state => state.User.userData.username,
+    isGroupAdmin: function () {
       if (this.user && this.group.GroupUsers) {
         const username = this.user.username;
         // hack - as long as they are not added (ie global admin) or added as an admin user.
         return !this.group.GroupUsers.some(user => username === user.username && !user.isAdmin);
       }
       return false;
-    },
-  },
+    }
+  }),
   methods: {
-    async removeUser(username) {
+    async removeUser(username, uiUser) {
+      if (username == uiUser) {
+        // TODO make this dialog look nicer (same for device)
+        var retVal = confirm("Are you sure you want to remove yourself from this group?  If you countinue " +
+        "will no longer be able to view recordings from the devices in this group and you will not be able to " +
+        "add yourself back to the group.\n\nDo you want to continue ?");
+        if( retVal == false ) {
+          return;
+        }
+      }
       await this.$store.dispatch('Groups/REMOVE_GROUP_USER', {userName: username, groupName: this.group.groupname});
     },
   }
@@ -127,7 +142,7 @@ export default {
   }
 
   @media (min-width: 992px) {
-    .users-detail && {
+    .users-detail {
       padding-right:30px;
     }
 
@@ -139,5 +154,12 @@ export default {
   h2 {
     font-size: large;
     margin-top: 2rem;
+  }
+
+  button.trash-button {
+    padding: 0;
+    background: inherit;
+    color: black;
+    border: none;
   }
 </style>
